requirements.txt
flask
gunicorn
requests
google-generativeai
Pillow
